+++ src/index.ts
@@ -1,10 +1,11 @@
 /**
  * API Node.js Client
  */
-import { service, transporters } from './abstract';
+import { factory, service, transporters } from './abstract';
 import ClientCredentialsClient from './auth/clientCredentialsClient';
 import OAuth2client from './auth/oAuth2client';
 
+export import factory = factory;
 export import service = service;
 export import transporters = transporters;